TG-150: Created AddTimesheet folder, renamed js and css files, renamed ModalS to WiderModal, amended bugon zip value of employee edit form
