+++ src/Components/Employee/Timesheet/timesheet.module.css
@@ -13,7 +13,3 @@
   justify-content: center;
 }
 
-.tableContainer{
-  width: 65%;
-  margin-bottom: 5%;
-}
